# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing incidents with AI classification",
        "title": "Incident Management API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/incidents/create": {
            "post": {
                "description": "Create a new incident with AI classification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "Create an incident",
                "parameters": [
                    {
                        "description": "Incident data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incidents.Incident"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/incidents.Incident"
                        }
                    }
                }
            }
        },
        "/incidents/get": {
            "get": {
                "description": "Get a list of all incidents",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "List incidents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/incidents.Incident"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "incidents.Incident": {
            "type": "object",
            "properties": {
                "affected_service": {
                    "type": "string"
                },
                "ai_category": {
                    "type": "string"
                },
                "ai_severity": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}